test: organize global search tests
